// types/product.ts
export interface Review {
  id: string;
  username: string;
  rating: number;
  description: string;
}

export interface Product {
  id: string;
  title: string;
  description: string;
  price: number;
  discountedPrice?: number; // optional hvis ikke alltid finnes
  image: {
    url: string;
    alt: string;
  };
  rating: number;
  tags?: string[];
  reviews?: Review[];
}